Remove /docs/ from /topics/ URLs
We've decided earlier that we want short URLs on the form

    https://valkey.io/topics/acl

The URLs (before this patch) contain /docs/topics/ and a trailing slash,
like this:

    https://valkey.io/docs/topics/acl/

This is to make the URLs shorter and more close to what's been decided.

In the docs repo, the links between the command pages and the topics
pages are on the form "../topics/acl.md" and "../commands/client-kill.md"
so by putting "topics" at the same level as "commands" on the website
too makes editing the docs more intuitive.

I can't get rid of the trailing slash right now, so with this patch,
the URLs look like

    https://valkey.io/topics/acl/

Redirects are added for the /docs/topics/ URLs.

Signed-off-by: Viktor Söderqvist <viktor.soderqvist@est.tech>