Adding analytics to valkey.io (#111)
### Description

Adding Fathom Analytics to site

Analytics are important as we make changes and add content to the site.
The data will enable us to find what pages perform well vs. those that
don't. It can also enable discovery of how users use the site itself,
which docs are more important to users, and a bevy of other useful data
as the project grows.

---------

Signed-off-by: Chris "Not So" Short <cbshort@amazon.com>
Signed-off-by: Chris "Not So" Short <chris@chrisshort.net>
Co-authored-by: Madelyn Olson <madelyneolson@gmail.com>