Add a permanent link for slack (#200)
### Description

Our slack was not publicly visible, and I'm not convinced the link will
be permanent. So add a redirect link on our page that you can use to
join the slack that will be permanent.
 
### Issues Resolved

Hopefully addresses
https://github.com/valkey-io/valkey-io.github.io/issues/198.

### Check List
- [X] Commits are signed per the DCO using `--signoff`

By submitting this pull request, I confirm that my contribution is made
under the terms of the BSD-3-Clause License.

Signed-off-by: Madelyn Olson <madelyneolson@gmail.com>