Implement a properly sorted downloads page (#201)
### Description

Added a quick and dirty sorting mechanism for sorting the downloads
page. Includes the latest version from each major, with a new sorting
based of a weight, which is a numeric representation of the semantic
version.

Notes that weights must be a positive number, so you have to do a
reverse sort and can't use a negative number for the version.

<img width="622" alt="image"
src="https://github.com/user-attachments/assets/6aafa534-a305-446c-8bc8-5b5aa4298fed"
/>
 
### Issues Resolved

<!-- List any issues this PR will resolve. -->
<!-- Example: closes #1234 -->

### Check List
- [x] Commits are signed per the DCO using `--signoff`

By submitting this pull request, I confirm that my contribution is made
under the terms of the BSD-3-Clause License.

---------

Signed-off-by: Madelyn Olson <madelyneolson@gmail.com>