Adds a clients overview page (#166)
### Description

Adds a client page to docs on the website. The page is added to the top
nav docs, called "Client Libraries", and its url is top level
`valkey.io/clients `.
This PR is dependent on the [complimentary docs
PR](https://github.com/valkey-io/valkey-doc/pull/164) being merged.

#### How it works - 
- The `init-topics.sh` script was modified to expect the path to the
clients docs folder as its second argument. It creates a symlink to it,
called `build-clients`. The clients folder in docs holds json files
specifying each client's details and features.
- `content/clients/_index.md` has in its frontmatter the list of paths
to the clients appearing in this page. This list is used by the
`client-list.html` template to render the specified clients' data. This
template also includes the `client-feature-table.html` template that
renders the same json files to produce a table comparing which features
each client implements.

### Issues Resolved
[#161](https://github.com/valkey-io/valkey-doc/issues/161)
Dependent on this PR being merged -
[#164](https://github.com/valkey-io/valkey-doc/pull/164)

### Check List
- [x] Commits are signed per the DCO using `--signoff`

By submitting this pull request, I confirm that my contribution is made
under the terms of the BSD-3-Clause License.

---------

Signed-off-by: lior sventitzky <liorsve@amazon.com>
Signed-off-by: Lior Sventitzky <liorsve@amazon.com>
Co-authored-by: Madelyn Olson <madelyneolson@gmail.com>