Adding functionality for the bloom module to have its commands displayed on the Valkey website (#212)
Related PR's
Bloom repo json command files:
https://github.com/valkey-io/valkey-bloom/pull/47
Valkey-doc repo: https://github.com/valkey-io/valkey-doc/pull/233

### Description
This PR will allow set the framework so that modules can have their
commands displayed on the valkey website (By adding the bloom module
commands in a way that can be easily expanded on). I have tried to make
this future proof by using a for loop on the `commands.html` page which
can be expanded by just adding any new folders we want to pull commands
from. For the `command-page.html` I have used an array to hold the data
from the multiple folders with commands and then get the first
occurrence that isn't empty (i.e the command belongs to that folder).
This will keep ability so that if the command doesn't exist we still
have the same fallback.

Updated the `init-commands.sh` to create a link for the bloom commands
as well and take in the bloom repository.

I have updated the README as well to include the new repo that will be
needed for the commands and the information change associated with now
expecting commands from the bloom repo.

Lastly updated the github workflow as well to also now build and take in
the bloom repo

**For screenshots of the new documentation the two pr's above
(https://github.com/valkey-io/valkey-doc/pull/233 and
https://github.com/valkey-io/valkey-bloom/pull/47) have screenshots of
all sections being added**

### Check List
- [x] Commits are signed per the DCO using `--signoff`

By submitting this pull request, I confirm that my contribution is made
under the terms of the BSD-3-Clause License.

Signed-off-by: zackcam <zackcam@amazon.com>