Correct versions on download page and template correction (#241)
### Description

downloads page did not show the correct version which were available.
Also corrects the template that will be used by the automation.

We are discontinuing the focal support as it s EOL

> Ubuntu 20.04 LTS runner will be removed on 2025-04-15. For more
details, see https://github.com/actions/runner-images/issues/11101

### Check List
- [x] Commits are signed per the DCO using `--signoff`

By submitting this pull request, I confirm that my contribution is made
under the terms of the BSD-3-Clause License.

Signed-off-by: Roshan Khatri <rvkhatri@amazon.com>