Adding implicit ACL categories to be displayed on website (#258)
### Description

Currently implicit categories aren't shown on the website. This pr
parses the command flags and adds the implicit categories based on this
Example images

![image](https://github.com/user-attachments/assets/26c29d51-4259-41bc-ab0f-61e03702c43d)

![image](https://github.com/user-attachments/assets/8fa6cfe2-a257-48bb-ac8c-ced4b1c1b7f1)

![image](https://github.com/user-attachments/assets/70284ade-63dc-401e-b332-d1847ddc305b)

If more screenshots are needed I can add to display all categories.

### Issues Resolved

This will fully finish
https://github.com/valkey-io/valkey-doc/issues/272#issuecomment-2864287464

### Check List
- [x] Commits are signed per the DCO using `--signoff`

By submitting this pull request, I confirm that my contribution is made
under the terms of the BSD-3-Clause License.

Signed-off-by: zackcam <zackcam@amazon.com>
Signed-off-by: Shai Zarka <zarkash@amazon.com>