Bug fix: allow local builds without requiring client directory (#243)
### Description

<!-- Describe what this change achieves-->

Currently building the zola site locally is not possible when the
clients directory is not built, as the `zola serve` command fails.
This is fixed by introducing an additional boolean variable
`clients_built` in the client's page frontmatter, which is set to false
by default, but updates to be true when running the
`init_topics_and_clients.sh` script successfully. The page loads only if
this var is set to true, otherwise a "Clients page not found" message is
displayed, similar to the behavior in the commands and topics
documentation.
 
### Issues Resolved

<!-- List any issues this PR will resolve. -->
<!-- Example: closes #1234 -->

### Check List
- [x] Commits are signed per the DCO using `--signoff`

By submitting this pull request, I confirm that my contribution is made
under the terms of the BSD-3-Clause License.

Signed-off-by: Lior Sventitzky <liorsve@amazon.com>