Changes variable names to prevent bizarre client container sidebar issue (#320)
### Description

It looks like in #299 the sidebar code was copy/pasted from the command
section into the command page. Both pages uses the global variable named
`command_data_obj`, so when there wasn't a container explicitly from the
command entry, it just used the one from the page.

In this PR, I namespaced the variable name to fix the command names in
the list. Additionally, I removed an unused global that sorted/grouped
all the commands.

I've also noticed that the build times have shot up after #299 to around
50 seconds - which is bonkers. Given the repetitious nature of this page
element, this should be abstracted into a macro and/or cached during
build. At time of writing there is 416 pages rendered in this section,
each one has to open 416 + 4 (json files that index all the commands).
That means that generating this sidebar adds 174,720 JSON file opens and
parses. I will file a seperate issue to track.

### Issues Resolved

#319

### Check List
- [x] Commits are signed per the DCO using `--signoff`

By submitting this pull request, I confirm that my contribution is made
under the terms of the BSD-3-Clause License.

Signed-off-by: Kyle J. Davis <kyledvs@amazon.com>