Adds atom template with logic for multiple RSS feeds (#327)
### Description

This pulls in source for the pre-built atom feed from zola, adds logic
to filter by "section", overrides the pre-built atom feed, and enables
an event-only feed.

In the end, the main atom feed (`/atom.xml`) is just blog posts. There
is now an additional atom feed for events section `/events/atom.xml`)
and for any future sections that enable the feed generaiton.


### Issues Resolved

#322

### Check List
- [x] Commits are signed per the DCO using `--signoff`

By submitting this pull request, I confirm that my contribution is made
under the terms of the BSD-3-Clause License.

---------

Signed-off-by: Kyle J. Davis <kyledvs@amazon.com>
Signed-off-by: Madelyn Olson <madelyneolson@gmail.com>
Co-authored-by: Madelyn Olson <madelyneolson@gmail.com>