generate once, load many (#330)
### Description

When the side bar was added to the command page it caused a massive
increase in build times (see #321). This was caused by each command page
needing to pull every command JSON file into to generate the list.

This PR takes a different approach, it adds a page (`/command-list/`),
that is generated once, which contains the all the commands formatted as
list items. Then, on each command page, it pulls that content in with
some simple javascript.

Full build times were previously taking 40-70 seconds. After this
change, the build times are taking approximately 1.2 seconds.

```
Building site...
Checking all internal links with anchors.
> Successfully checked 0 internal link(s) with anchors.
-> Creating 564 pages (0 orphan) and 8 sections
Warning: Highlight language plaintext not found
Done in 1.2s.
```
 
### Issues Resolved

#321

### Check List
- [x] Commits are signed per the DCO using `--signoff`

By submitting this pull request, I confirm that my contribution is made
under the terms of the BSD-3-Clause License.

Signed-off-by: Kyle J. Davis <kyledvs@amazon.com>