change filename to denote rc1 (#336)
### Description

As per #211, this will make sure that there is a release page for `rc1`
distinct from GA and other release candidates.
### Issues Resolved

(this doesn't solve 211 - that will need to be fixed in the release
automation)

### Check List
- [x] Commits are signed per the DCO using `--signoff`

By submitting this pull request, I confirm that my contribution is made
under the terms of the BSD-3-Clause License.

Signed-off-by: Kyle J. Davis <kyledvs@amazon.com>