fixes title metadata that messes with LinkedIn (#358)
### Description

On LinkedIn, sharing a post from Valkey.io it always makes the title
weirdly "command reference".

Looking into the LinkedIn Debugging tool it seems like it can't infer
our title so we need to add an explicit `og:title` metadata tag. This
adds the same logic as the title with colon seperating the page type and
title (instead of the middot entity, which is not allowed in this
format).


### Issues Resolved

n/a

### Check List
- [x] Commits are signed per the DCO using `--signoff`

By submitting this pull request, I confirm that my contribution is made
under the terms of the BSD-3-Clause License.

Signed-off-by: Kyle J. Davis <kyledvs@amazon.com>