Bitnami post update after changes in their catalog (#360)
### Description

This posts adds a noticed that the post for the bitnami helm chart is
out of date. I opted to add a notice instead of deleting so readers
could be directed to work on the replacement.
 
### Issues Resolved

#355

### Check List
- [x] Commits are signed per the DCO using `--signoff`

By submitting this pull request, I confirm that my contribution is made
under the terms of the BSD-3-Clause License.

---------

Signed-off-by: Kyle J. Davis <kyledvs@amazon.com>