Community event landing page  (#370)
### Description

(exactly the same as #364 but based on commit 4888f10... for whatever
reason GitHub _did not_ like the combinations of changes - it was far
easier to just manually move these changes over top of 4888f10 and do
another PR)

This PR refactors the /community/ page into a section. It allows us to
add unique landing page variations of the community page.

For example, if you add an markdown file to /content/community/ this
will create a new URL under /community/.

This is especially useful for events where we may want to direct
attendees to a specific set of resources for that event. For example, I
could send PyConAfrica attendees to a vakey.io/community/pycon-africa
via a QR code. On this specific landing page it would contain
information about demos and/or slides relevant to the event along side
our standard items on the community page.
 
### Issues Resolved

n/a

### Check List
- [x] Commits are signed per the DCO using `--signoff`

By submitting this pull request, I confirm that my contribution is made
under the terms of the BSD-3-Clause License.

Signed-off-by: Kyle J. Davis <kyledvs@amazon.com>