Try Valkey for Workshops (#378)
### Description

This PR accomplishes two things:
1. Generalized the try-valkey interface allowing for multiple, isolated
instances of try-valkey to co-exist on valkey.io
2. Adds a new instance of try-valkey for the PyCon Africa conference
workshop.
 
Note: the actual distribution needs to be uploaded to the
download.valkey.io CDN for the workshop to be functional.

### Issues Resolved

n/a

### Check List
- [x] Commits are signed per the DCO using `--signoff`

By submitting this pull request, I confirm that my contribution is made
under the terms of the BSD-3-Clause License.

---------

Signed-off-by: Kyle J. Davis <kyledvs@amazon.com>