Add Scarf pixel (#350)
### Description

As discussed with @pnbrown today, this PR adds a Scarf pixel to the
footer of the site for analytics purposes in a way that's compliant with
LF policies.

Insights will be available to any maintainers in the Valkey org on
https://scarf.sh.

### Issues Resolved

None. 

### Check List
- [x] Commits are signed per the DCO using `--signoff`

By submitting this pull request, I confirm that my contribution is made
under the terms of the BSD-3-Clause License.

Signed-off-by: aviaviavi <avipress@gmail.com>