adds custom meta to head via front matter (#394)
### Description

In #380, @dmolik was trying to add a custom meta tags for the go package
manager. His solution was limited and specific to to the specific
situation with the go package manager but custom tags in the header are
not an uncommon ask, so I wanted a generalizable solution. This PR adds
the ability for add tags directly to any page in the meta block.

 
### Issues Resolved

n/a

### Check List
- [x] Commits are signed per the DCO using `--signoff`

By submitting this pull request, I confirm that my contribution is made
under the terms of the BSD-3-Clause License.

Signed-off-by: Kyle J. Davis <kyledvs@amazon.com>