adds link to event registration and modifies template for footer (#415)
### Description

1. Adds the link to the event registration (in two langauges)
2. Modifies the template to allow for footer injection. In this case, we
need to put the credit for the image in the header, but it can be
generally useful for anything we don't want to end up on the page
itself.
 
### Issues Resolved

n/a

### Check List
- [x] Commits are signed per the DCO using `--signoff`

By submitting this pull request, I confirm that my contribution is made
under the terms of the BSD-3-Clause License.

Signed-off-by: Kyle J. Davis <kyledvs@amazon.com>